mcp/memory_python/src/pyproject.toml
[project]
name = "mcp_memory_python"
version = "0.0.1"
description = "An MCP server project to manage memory backed by MongoDB"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "mcp>=1.2.0",
 "pymongo>=4.10.1",
 "pytest>=8.3.4",
 "python-dotenv>=1.0.1",
]
[[project.authors]]
name = "Tim"
email = ""

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp_memory_python = "mcp_memory_python:main"
